[libc] Add separate `rand` implementation for baremetal
Summary:
Baremetal targets likely do now want multi-threaded support. Split the
implementation out to make it separate.
